[NT-0] feat: update tests to floats

+++ Tests/src/PublicAPITests/ComponentTests/ConversationComponentTests.cpp
@@ -1584,7 +1584,7 @@ CSP_PUBLIC_TEST(CSPEngine, ConversationTests, ConversationComponentCreateAnnotat
 
     static const csp::common::Vector3 TestConversationAuthorCameraPosition { 1.f, 2.f, 3.f };
     static const csp::common::Vector4 TestConversationAuthorCameraRotation { 4.f, 5.f, 6.f, 7.f };
-    static constexpr const uint16_t TestConversationFov = 90;
+    static constexpr const float TestConversationFov = 90.f;
 
     std::vector<char> TestAnnotationDataV = PngHeader;
     TestAnnotationDataV.push_back('1');
@@ -1699,7 +1699,7 @@ CSP_PUBLIC_TEST(CSPEngine, ConversationTests, ConversationComponentCreateAnnotat
 
     static const csp::common::Vector3 TestAuthorCameraPosition { 8.f, 9.f, 10.f };
     static const csp::common::Vector4 TestAuthorCameraRotation { 11.f, 12.f, 13.f, 14.f };
-    static constexpr const uint16_t TestFov = 100;
+    static constexpr const float TestFov = 100.f;
 
     std::vector<char> TestAnnotationData2V = PngHeader;
     TestAnnotationData2V.push_back('3');
@@ -2092,7 +2092,7 @@ CSP_PUBLIC_TEST(CSPEngine, ConversationTests, ConversationComponentOverwriteAnno
     {
         static const csp::common::Vector3 TestAuthorCameraPosition { 1.f, 2.f, 3.f };
         static const csp::common::Vector4 TestAuthorCameraRotation { 4.f, 5.f, 6.f, 7.f };
-        static constexpr const uint16_t TestFov = 90;
+        static constexpr const float TestFov = 90.f;
         std::vector<char> TestAnnotationDataV = PngHeader;
         TestAnnotationDataV.push_back('1');
         std::vector<char> TestAnnotationThumbnailDataV = PngHeader;
@@ -2122,7 +2122,7 @@ CSP_PUBLIC_TEST(CSPEngine, ConversationTests, ConversationComponentOverwriteAnno
     {
         static const csp::common::Vector3 TestAuthorCameraPosition { 8.f, 9.f, 10.f };
         static const csp::common::Vector4 TestAuthorCameraRotation { 11.f, 12.f, 13.f, 14.f };
-        static constexpr const uint16_t TestFov = 100;
+        static constexpr const float TestFov = 100.f;
         bool AssetOverwriteLogCalled = false;
 
         std::vector<char> TestAnnotationDataV = PngHeader;
